+++ src/main/java/com/flab/sooldama/domain/product/service/ProductService.java
@@ -16,7 +16,7 @@ public class ProductService {
         List<Product> products = productMapper.selectProducts(offset, limit);
         return ProductsResponse.from(products);
     }
-    public ProductsResponse getProductsByCategoryId(Integer offset, Integer limit, Integer categoryId) {
+    public ProductsResponse getProductsByCategoryId(Integer offset, Integer limit, Long categoryId) {
         List<Product> products = productMapper.selectProductsByCategoryId(offset, limit, categoryId);
         return ProductsResponse.from(products);
     }
